Split amp-base-carousel component files
