+++ extensions/amp-base-carousel/1.0/test-e2e/test-arrows-non-looping.js
@@ -20,7 +20,7 @@ import {
   getPrevArrow,
   getScrollingElement,
 } from './helpers';
-import {useStyles} from '../base-carousel.jss';
+import {useStyles} from '../component.jss';
 
 /** The total number of slides in the carousel */
 const SLIDE_COUNT = 4;
