+++ extensions/amp-base-carousel/1.0/test-e2e/test-mixed-lengths-no-snap.js
@@ -15,7 +15,7 @@
  */
 
 import {getCarousel, getScrollingElement, getSlide} from './helpers';
-import {useStyles} from '../base-carousel.jss';
+import {useStyles} from '../component.jss';
 
 const pageWidth = 800;
 const pageHeight = 600;
