+++ extensions/amp-base-carousel/1.0/test-e2e/test-vertical.js
@@ -15,7 +15,7 @@
  */
 
 import {getCarousel, getScrollingElement, getSlide} from './helpers';
-import {useStyles} from '../base-carousel.jss';
+import {useStyles} from '../component.jss';
 
 const pageWidth = 600;
 const pageHeight = 600;
