+++ extensions/amp-base-carousel/1.0/test/test-amp-base-carousel.js
@@ -24,7 +24,7 @@ import {mod} from '../../../../src/utils/math';
 import {setStyles} from '../../../../src/style';
 import {toArray} from '../../../../src/types';
 import {toggleExperiment} from '../../../../src/experiments';
-import {useStyles} from '../base-carousel.jss';
+import {useStyles} from '../component.jss';
 import {waitFor, whenCalled} from '../../../../testing/test-helper';
 
 describes.realWin(
