+++ extensions/amp-inline-gallery/1.0/storybook/Basic.js
@@ -15,7 +15,7 @@
  */
 
 import * as Preact from '../../../../src/preact';
-import {BaseCarousel} from '../../../amp-base-carousel/1.0/base-carousel';
+import {BaseCarousel} from '../../../amp-base-carousel/1.0/component';
 import {InlineGallery} from '../inline-gallery';
 import {Pagination} from '../pagination';
 import {Thumbnails} from '../thumbnails';
